requirements.txt
fastapi
uvicorn
pillow
numpy
keras
torch
torchvision
tensorflow
python-multipart
